requirements.txt
zeep
djangorestframework
